feat: move logic to useGetActivePools hook
